Feature #2: Cron job to purge old reports in cyclic way. Using ENV['PUPPET_HERALD_PURGE_CRON'] - by default '*/30 * * * *' and ENV['PUPPET_HERALD_PURGE_LIMIT'] - by default '90d'

+++ lib/puppet-herald/models/node.rb
@@ -1,5 +1,6 @@
 require 'puppet-herald/models'
 require 'puppet-herald/models/report'
+require 'sinatra/activerecord'
 
 # A module for Herald
 module PuppetHerald
@@ -35,9 +36,9 @@ module PuppetHerald
       #
       # @return [Integer] number of empty node deleted
       def self.delete_empty
-        joins('LEFT JOIN "reports" ON "reports"."node_id" = "nodes"."id"')
-          .where('"reports"."node_id" IS NULL')
-          .delete_all
+        joinsql = 'LEFT JOIN "reports" ON "reports"."node_id" = "nodes"."id"'
+        wheresql = '"reports"."node_id" IS NULL'
+        joins(joinsql).where(wheresql).delete_all if joins(joinsql).where(wheresql).count > 0
       end
 
       # Gets a node with prefetched reports
